requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.5.0
streamlit>=1.20.0
filterpy>=1.4.5
scipy>=1.7.0
Pillow>=9.0.0

